Move Formula types to formula.ts
