+++ server/src/formulas/index.ts
@@ -4,8 +4,7 @@
 
 import * as yup from 'yup';
 
-import { Formula } from '../types/attainable';
-import { CalculationResult, registerFormula, Status } from '../types/formulas';
+import { CalculationResult, Formula, registerFormula, Status } from '../types/formulas';
 
 registerFormula(
   Formula.Manual,
