+++ server/src/types/attainable.ts
@@ -23,12 +23,3 @@ export interface AttainableData {
   subAttainments?: Array<AttainableData>
 }
 
-export enum Formula {
-  Manual = 'MANUAL',
-  WeightedAverage = 'WEIGHTED_AVERAGE',
-}
-
-export interface FormulaParams {
-  min: number;
-  max: number;
-}
